Show user the download costs. Some improvements.
Show user the job output files download cost.
Get the job output details via /jobs/' + job.id + '/outputFiles?projection=detailedFtepFile'.
Fix job wms layers toggling.
When a new item is created, select it (cannot expect it to be the last one).
Ask confirmation when deleting a service.
Set the jobs filter order to same as it is for others.

+++ f-tep-portal/src/main/resources/app/scripts/controllers/developer/service.js
@@ -113,9 +113,14 @@ define(['../../ftepmodules'], function (ftepmodules) {
             });
         };
 
-        $scope.removeService = function(service){
-            ProductService.removeService(service).then(function(){
-                ProductService.refreshServices('development', 'Remove', service);
+        $scope.removeService = function(event, service){
+            CommonService.confirm(event, 'Are you sure you want to delete this service: "' + service.name + '"?').then(function (confirmed){
+                if(confirmed === false){
+                    return;
+                }
+                ProductService.removeService(service).then(function(){
+                    ProductService.refreshServices('development', 'Remove', service);
+                });
             });
         };
 
@@ -124,7 +129,7 @@ define(['../../ftepmodules'], function (ftepmodules) {
 
                     $scope.createService = function () {
                         ProductService.createService($scope.newItem.name, $scope.newItem.description, $scope.newItem.title).then(function (newService) {
-                            ProductService.refreshServices('development', 'Create');
+                            ProductService.refreshServices('development', 'Create', newService);
                         });
                         $mdDialog.hide();
                     };
